+++ circular_buffer.h
@@ -5,27 +5,27 @@
 
 typedef struct
 {
-    int *ptrBuffer;
-    int head;
-    int tail;
-    int size;
-    int count;
+    uint32_t *ptrBuffer;
+    uint32_t head;
+    uint32_t tail;
+    uint32_t size;
+    uint32_t count;
 } CircularBuffer;
 
-void arrangeCircularBuffer(CircularBuffer *ptr, int *buffer, int size);
+void arrangeCircularBuffer(CircularBuffer *ptr, uint8_t *buffer, uint32_t size);
 
-void pushItemToBuff(CircularBuffer *ptr, int item);
+void pushItemToBuff(CircularBuffer *ptr, uint8_t item);
 
-_Bool popItemFromBuff(CircularBuffer *ptr, int *item);
+_Bool popItemFromBuff(CircularBuffer *ptr, uint8_t *item);
 
-_Bool pickItemFromBuff(CircularBuffer *ptr, int *item, int offset);
+_Bool pickItemFromBuff(CircularBuffer *ptr, uint8_t *item, uint32_t offset);
 
 _Bool isBufferEmpty(CircularBuffer *ptr);
 
 void deleteFromHead(CircularBuffer *ptr);
 
-int getHead(CircularBuffer *ptr);
+uint32_t getHead(CircularBuffer *ptr);
 
-int getCount(CircularBuffer *ptr);
+uint32_t getCount(CircularBuffer *ptr);
 
 #endif
